+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1 - Second Release
+* Changed name from Encourager to Rubber Ducky
+* Added responses feature, which is the core function of rubber ducky
+
 ## 0.1.0 - First Release
 * Every feature added
 * Every bug fixed